requirements.txt
asgiref==3.5.2
attrs==22.1.0
Automat==22.10.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
cssselect==1.2.0
Django==4.0
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
drf-yasg==1.21.4
filelock==3.8.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itemadapter==0.7.0
itemloaders==1.0.6
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
lxml==4.9.1
MarkupSafe==2.1.1
mysqlclient==2.1.1
packaging==21.3
parsel==1.7.0
Pillow==9.3.0
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.6
pyOpenSSL==22.1.0
pyparsing==3.0.9
pytz==2022.6
PyYAML==6.0
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
Scrapy==2.7.1
scrapy-djangoitem==1.1.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
tldextract==3.4.0
Twisted==22.10.0
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
w3lib==2.1.0
zope.interface==5.5.2
